Added user class as requested
